Make DIR be faithfull to the *right-window* spec
Unfortunately, dcd62ea7f333d799e60a5161d401704eb34d944b broke the DIR
descriptor.

Example:

+-------+-------+
|       | b | c |
|   a   |-------|
|   ^   |   d   |
+-------+-------+

If `a` is focused and `d` is the most recently focused node, `east` will
resolve to `b` instead of `d`, because the center `b` is closer to the
center of `a` than `d`.

Another problem is that, given a rectangle (x, y, w, h), the point
(x+w, y+h) was considered to be inside of it: this created false
positives in on_dir_side.

Fixes #594.